cleanup: split BlockPreviewViews.swift into focused files by block type group
BlockPreviewViews.swift (1124 lines) split into 6 files:
- BlockPreviewHelpers.swift: shared helpers (hexColor, markdownText, statusColor)
- BlockPreviewInteractive.swift: Confirmation, Prompt, ChatPrompt, Picker
- BlockPreviewStatus.swift: Status, Alert, InfoCard, Countdown
- BlockPreviewContent.swift: List, Detail, IconCard, FeedItem, ClaudeMessage
- BlockPreviewSession.swift: AgentSession, RepoRow, StartSession, Diagnostics
- BlockPreviewViews.swift: dispatcher only (~42 lines)
